+++ pom.xml
@@ -104,7 +104,7 @@
     <module>lens</module>
     <module>cassandra</module>
     <module>elasticsearch</module>
-    <module>tachyon</module>
+    <module>alluxio</module>
     <module>zeppelin-web</module>
     <module>zeppelin-server</module>
     <module>zeppelin-distribution</module>
